Wsei-Backend-Kursy/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==2.5.1
passlib==1.7.4
psycopg2==2.9.6
SQLAlchemy==2.0.17
requests==2.31.0
python-dotenv==1.0.0
Flask-JWT-Extended==4.5.2
flask-restx==1.1.0
flask-cors==3.0.10